Java project 1 update

+++ JAVA/src/com/company/Point2D.java
@@ -1,9 +1,9 @@
 package com.company;
 
 public class Point2D {
-    protected Double x, y;
+    protected Integer x, y;
 
-    Point2D(Double x_val, Double y_val) {
+    Point2D(Integer x_val, Integer y_val) {
         x = x_val;
         y = y_val;
     }
